test: mapped and reenabled pool edge cases
